+++ src/Core/Adapter/AdapterFactory/AdapterFactory.php
@@ -15,6 +15,7 @@ namespace Core\Adapter\AdapterFactory;
 
 
 use Core\Adapter\AdapterInterface;
+use Core\Cache\CacheInterface;
 use Core\Entity\EntityInterface;
 use Core\Entity\Error\EntityAdapterConfigException;
 use Core\Entity\Error\EntityClassNotExistsException;
@@ -24,14 +25,19 @@ class AdapterFactory implements AdapterFactoryInterface
 {
     /** @var ServiceLocatorInterface */
     protected $serviceLocator;
+    /** @var null|CacheInterface */
+    protected $cacheAdapter = null;
 
     /**
      * AdapterFactory constructor.
      * @param ServiceLocatorInterface $serviceLocator
+     * @param CacheInterface|null $baseCacheAdapter
+     * @param null $fallBackCache
      */
-    public function __construct(ServiceLocatorInterface $serviceLocator)
+    public function __construct(ServiceLocatorInterface $serviceLocator, CacheInterface $baseCacheAdapter = null)
     {
         $this->serviceLocator = $serviceLocator;
+        $this->cacheAdapter = $baseCacheAdapter;
     }
 
 
@@ -57,6 +63,19 @@ class AdapterFactory implements AdapterFactoryInterface
         /** @var AdapterInterface $adapter */
         $adapter = $this->serviceLocator->get($adapter);
         $adapter->setAdapterFactory($this);
+
+        if (method_exists($adapter, "setCache")) {
+            $cache = $this->cacheAdapter;
+
+            if (!empty($entityConfiguration["cache"])) {
+                /** @var CacheInterface $cache */
+                $cache = new $entityConfiguration["cache"]["main"];
+                $fallBackCache = new $entityConfiguration["cache"]["fallback"];
+                $cache->setFallBack($fallBackCache);
+            }
+
+            $adapter->setCache($cache);
+        }
 
         return $adapter;