+++ src/Core/Cache/CacheFactory.php
@@ -1,0 +1,32 @@
+<?php
+/**
+ * <strong>Name :  CacheFactory.php</strong></br>
+ * <strong>Desc :  [Put a description here]</strong></br>
+ *
+ * @category rdsdb-api
+ * @package Core\Cache
+ * @author Juan Pablo Cruz Maseda <pablo.cruz@digimobil.es>
+ * @copyright 2019 digimobil.es
+ * @link https://gq-api.digimobil.es
+ */
+
+
+namespace Core\Cache;
+
+
+use Zend\ServiceManager\FactoryInterface;
+use Zend\ServiceManager\ServiceLocatorInterface;
+
+class CacheFactory implements FactoryInterface
+{
+    /**
+     * @inheritDoc
+     */
+    public function createService(ServiceLocatorInterface $serviceLocator)
+    {
+        $adapter = $serviceLocator->get("Core\CacheAPCAdapter");
+        return new BaseCache($adapter);
+    }
+
+
+}