+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # Changelog
+
+- Implement cache with APCu, Array and Memcached/Couchebase adapters.